+++ e-shop/src/main/java/org/ehais/shop/service/HaiGoodsService.java
@@ -1,7 +1,11 @@
 package org.ehais.shop.service;
 
 import javax.servlet.http.HttpServletRequest;
+
 import org.ehais.service.CommonService;
+import org.ehais.shop.model.HaiGoods;
+import org.ehais.shop.model.HaiGoodsAttr;
+import org.ehais.shop.model.HaiGoodsWithBLOBs;
 import org.ehais.tools.EConditionObject;
 import org.ehais.tools.ReturnObject;
 
@@ -10,11 +14,11 @@ public interface HaiGoodsService extends CommonService{
 	public ReturnObject<HaiGoods> goods_list_json(HttpServletRequest request,EConditionObject condition,Integer goodsAttrId,String goodsName) throws Exception;
 	public ReturnObject<HaiGoodsWithBLOBs> goods_insert(HttpServletRequest request) throws Exception;
 	public ReturnObject<HaiGoodsWithBLOBs> goods_insert_submit(HttpServletRequest request,HaiGoodsWithBLOBs model) throws Exception;
-	public ReturnObject<HaiGoodsWithBLOBs> goods_update(HttpServletRequest request,Integer goodsId) throws Exception;
+	public ReturnObject<HaiGoodsWithBLOBs> goods_update(HttpServletRequest request,Long goodsId) throws Exception;
 	public ReturnObject<HaiGoodsWithBLOBs> goods_update_submit(HttpServletRequest request,HaiGoodsWithBLOBs model) throws Exception;
-	public ReturnObject<HaiGoodsWithBLOBs> goods_info(HttpServletRequest request,Integer goodsId) throws Exception;
-	public ReturnObject<HaiGoodsWithBLOBs> goods_find(HttpServletRequest request,Integer goodsId) throws Exception;
-	public ReturnObject<HaiGoods> goods_delete(HttpServletRequest request,Integer goodsId) throws Exception;
+	public ReturnObject<HaiGoodsWithBLOBs> goods_info(HttpServletRequest request,Long goodsId) throws Exception;
+	public ReturnObject<HaiGoodsWithBLOBs> goods_find(HttpServletRequest request,Long goodsId) throws Exception;
+	public ReturnObject<HaiGoods> goods_delete(HttpServletRequest request,Long goodsId) throws Exception;
 	
 
 ////////////////////////////////////////////////////////////////////
